moved to jekyll
Wow. That was a lot of work. Now I'm starting to work with jekyll. And set up a basic blog page (that's not really functional yet) with sample content (that is not my portfolio). Trying to improve on my bootstrap knowledge, reading the documentation and building as I go.
I hope to continue working on this over the next month, and hopefully have a complete thing by the end of the year.